Added require_components macro in macros.cmake, and processing in tools.cmake.
In project.cmake, ideally, we should be able to do require_components(i2c driver)...etc to get the required components. This is not board-specific. Theoretically, this should be usable in STM32 and in others. Although, in different boards' common files or private files, there should be a tools.cmake or any other cmake file that will process this. In my tools.cmake that I added in the ESP board folder, I implemented a system where it gets the CMakeLists.txt of each component provided by the esp-idf. All we have to do is specify the name of the component, and it will get it. I have to override the idf_component_register so that the added files in CMakeLists.txt will be added to the correct variables so that the ECF will recognise and will properly compile them.